Project V.I.D -Working Developer Build-Alpha-1.0.9
Build's .gitIgnore file finally works as intended and the Test Environment is now ready. The fix was moving the Test Environment into a folder and telling git to ignore that folder instead of the scene files themselves!

Furthermore edited the build settings to have the other scenes work as well. The Repo is finally equipped and ready for this project's development.