Project V.I.D -Flashlights&Bots-1.0.12
Added a rough flashlight prototype that turns on/off with a click of the index trigger. Also added a proof of concept for the bot "A.I." that'll make their way toward the player. So far, the bots only used Unity's "NavMesh Agent" that essentially works off of the A* algorithm.

Note:
- Updated the "MainRoom" Scene" to the current VR-Environment prototype scene.
- Made a script for the Flashlight prefab to alter/change it's flashlight strengths
- Added different fog colors/effects in the "Fog Emitter" prefab.