Project V.I.D -THEBLACKSMOGISALIVE-1.0.14
Flash it once, it USUALLY goes back to it's spawn point. Flash it twice, It MAY get angry and charge at you. Shake the flashlight on it? I have no idea what it'll try next! Exactly like a distracted driver!

This monster has a mind of it's own. It took forever to program it and it would NOT behave the way I intend it to... But it works? I suspect it's a Coroutine bug... But holy damn is it almost sentient?