Project V.I.D -FlashlightChangeForSmogMonster-1.0.25
Fixed the bug where the smog monster wasn't reacting to the flashlight correctly. Apparently, due to the flashlight toggle not always keeping the flashlight on, the flashlight should instead draw the ray if it's on and not only during the toggle.

Moved that section of code and that fixed the interaction!